requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.153
newspaper3k==0.2.8
nltk==3.8.1
python-dotenv==1.0.1
Requests==2.32.3
textblob==0.18.0.post0
vaderSentiment==3.3.2
vaderSentiment==3.3.2
